Avoid ReactMount warnings
The situation where a WinJS control manages child WinJS controls
and moves them around in the DOM is tricky. Hub and HubSection
are an example of this. The problem is that the child WinJS
controls are described as React components by the app. We'd like
to mount those React components in the DOM but React doesn't
like how WinJS is manipulating the DOM. Specifically:
  - WinJS moves around the winControl.element. In react-winjs,
    components generate winControl.element in their renderers.
    Consequently, we need to call React.render on the parent.
    However, WinJS pulls winControl.element out of parent and moves
    it around the DOM.
  - After the component initially renders and the winControl is
    instantiated, the winControl may completely replace its innerHTML.
    Consequently, on subsequent calls to React.render, React does not
    believe this DOM is under its control so it creates a new
    component rather than updating an existing one.

The previous solution rendered WinJSChildComponents exactly 1 time
in an unparented element with React.render and then gave the
results to the parent winControl. When the parent winControl
unparented the component DOM node from the element it was mounted
in, React generated this warning:
  ReactMount: Root element has been removed from its original
  container. New container:

The new solution is cleaner. I'd still like to find a cleaner
solution but at least this one doesn't generate any warnings.

The new solution essentially gives up on trying to make
WinJSChildComponents React components. Instead, it just
uses React components for their data to figure out how to
update the element it shares with the WinJS control. It uses
React.renderToStaticMarkup for the initial render so that it
can generate the DOM for the component and give it to the
winControl to own without React mounting it and getting upset
when it loses control of the DOM.

In addition to doing this solution, I also added some console
warnings/errors to point out when apps violate the
restrictions of a WinJSChildComponent:
  - key is required
  - ref doesn't work