Initialize className and style before winControl instantiation
This is important when className or style may affect the
measurements of the winControl.

This new behavior is then used to initialize an example FlipView
to a width of 200px. Prior to this change to react-winjs, the
FlipView would have measured itself when its width was the width
of its parent.